monte_carlo_gradients/requirements.txt
absl-py==0.7.1
dm-sonnet==1.34
numpy==1.16.4
scipy==1.4.1
scikit-learn==0.21
tensorflow==1.15.2
tfp-nightly==0.9.0.dev20191101